Jobs: Cast return value of 'jobswp_days_until_pruning' filter as int.
git-svn-id: https://meta.svn.wordpress.org/sites/trunk@8405 74240141-8908-4e6f-9713-ba540dce6ec7

+++ jobs.wordpress.net/public_html/wp-content/plugins/jobswp/jobswp.php
@@ -68,7 +68,7 @@ class Jobs_Dot_WP {
 	 * @return int
 	 */
 	public static function get_days_until_pruning() {
-		return apply_filters( 'jobswp_days_until_pruning', 21 );
+		return (int) apply_filters( 'jobswp_days_until_pruning', 21 );
 	}
 
 	/**
